Split code into api, swagger, renderers, runner and plugin.
