+++ generator-api/src/main/scala/com/lightbend/lagom/spec/ResourceUtils.scala
