+++ generator-api/src/main/scala/com/lightbend/lagom/spec/model/DescriptorModelling.scala
