+++ generator-api/src/main/scala/com/lightbend/lagom/spec/render/LagomRender.scala
@@ -1,0 +1,7 @@
+package com.lightbend.lagom.spec.render
+
+import com.lightbend.lagom.spec.LagomGeneratorTypes
+
+trait LagomRender {
+  val render: LagomGeneratorTypes.Render
+}
