+++ generator-api/src/main/scala/com/lightbend/lagom/spec/render/ModelRender.scala
@@ -1,0 +1,7 @@
+package com.lightbend.lagom.spec.render
+
+import com.lightbend.lagom.spec.LagomGeneratorTypes
+
+trait ModelRender {
+  def render: LagomGeneratorTypes.ModelRender
+}
