+++ lagom-renderers/javadsl/src/test/scala/com/lightbend/lagom/spec/render/JavaLagomDescriptorRenderSpec.scala
