+++ runner/src/it/resources/java-custom-model-generator-spec/StatusEnum.java.txt
