+++ runner/src/it/resources/java-custom-model-generator-spec/swagger.json
