+++ runner/src/it/resources/lagom-generator-spec/sample-java-descriptor.txt
