+++ runner/src/it/resources/lagom-generator-spec/swagger.json
