+++ runner/src/it/scala/com/lightbend/lagom/spec/JavaCustomModelGeneratorSpec.scala
