+++ spec-parsers/openapi/src/test/resources/openapi-v2-parser/swagger.json
