+++ spec-parsers/openapi/src/test/scala/com/lightbend/lagom/spec/parser/OpenApiV2ParserSpec.scala
