--- src/main/scala/com/lightbend/lagom/spec/render/LagomRender.scala
@@ -1,36 +1,0 @@
-package com.lightbend.lagom.spec.render
-
-import java.io.File
-
-import com.lightbend.lagom.spec.LagomGeneratorTypes
-import com.lightbend.lagom.spec.LagomGeneratorTypes.{ GeneratedCode, Output }
-import com.lightbend.lagom.spec.model.Service
-import com.lightbend.lagom.spec.render.descriptor.JavaLagomDescriptorRender
-import com.lightbend.lagom.spec.render.model.JavaModelRender
-
-/**
- *
- */
-trait LagomRender {
-  val render: LagomGeneratorTypes.Render
-}
-
-object JavaLagomRender {
-
-  val render: LagomGeneratorTypes.Render = { service =>
-    val descriptorFileName = getPath(service, s"${service.interfaceName}.java")
-    val descriptor: String = JavaLagomDescriptorRender.render(service)
-
-    val customModels = service.customModels.map { model =>
-      val content = JavaModelRender.render(service, model)
-      val code = GeneratedCode(getPath(service, model.className), content)
-      (model.className, code)
-    }.toMap
-
-    Output(GeneratedCode(descriptorFileName, descriptor), customModels)
-  }
-
-  private def getPath(service: Service, className: String): String = {
-    s"${service.`package`.replaceAll("\\.", File.separator)}${File.separator}${className}.java"
-  }
-}
